requirements.txt
Flask==3.0.0
Flask-CORS==4.0.1
mysql-connector-python==9.0.0
tensorflow==2.17.0
Pillow==10.4.0
numpy==1.26.4
requests==2.32.3
gunicorn==22.0.0
